refactor: split map stories
