+++ packages/ui/core-components/src/lib/unsorted/viz/map/AreaMap.stories.svelte
@@ -1,0 +1,39 @@
+<script context="module">
+	/** @type {import("@storybook/svelte").Meta}*/
+	export const meta = {
+		title: 'Charts/Maps/AreaMap',
+		component: AreaMap
+	};
+</script>
+
+<script>
+	import { Story } from '@storybook/addon-svelte-csf';
+	import { Query } from '@evidence-dev/sdk/usql';
+	import { query } from '@evidence-dev/universal-sql/client-duckdb';
+
+	import AreaMap from './AreaMap.svelte';
+
+	/** @type {typeof query} */
+	const slowQuery = async (...args) => {
+		await new Promise((resolve) => setTimeout(resolve, 5_000));
+		return query(...args);
+	};
+
+	const la_zip_sales = Query.create(
+		`select * from la_zip_sales where zip_code <> 90704 order by 1`,
+		query
+	);
+
+	const slow_la_zip_sales = Query.create(
+		`select * from la_zip_sales where zip_code <> 90704 order by 1 limit 100`,
+		slowQuery
+	);
+</script>
+
+<Story name="Basic Usage" parameters={{ chromatic: { disableSnapshot: true } }}>
+	<AreaMap data={la_zip_sales} geoId="ZCTA5CE10" value="sales" areaCol="zip_code" />
+</Story>
+
+<Story name="Loading" parameters={{ chromatic: { disableSnapshot: true } }}>
+	<AreaMap data={slow_la_zip_sales} geoId="ZCTA5CE10" value="sales" areaCol="zip_code" />
+</Story>
