+++ packages/ui/core-components/src/lib/unsorted/viz/map/BubbleMap.stories.svelte
@@ -1,0 +1,36 @@
+<script context="module">
+	/** @type {import("@storybook/svelte").Meta}*/
+	export const meta = {
+		title: 'Charts/Maps/BubbleMap',
+		component: BubbleMap
+	};
+</script>
+
+<script>
+	import { Story } from '@storybook/addon-svelte-csf';
+	import { Query } from '@evidence-dev/sdk/usql';
+	import { query } from '@evidence-dev/universal-sql/client-duckdb';
+
+	import BubbleMap from './BubbleMap.svelte';
+
+	/** @type {typeof query} */
+	const slowQuery = async (...args) => {
+		await new Promise((resolve) => setTimeout(resolve, 5_000));
+		return query(...args);
+	};
+
+	const la_locations = Query.create(`select * from la_locations order by 1`, query);
+
+	const slow_la_locations = Query.create(
+		`select * from la_locations order by 1 limit 100`,
+		slowQuery
+	);
+</script>
+
+<Story name="Basic Usage" parameters={{ chromatic: { disableSnapshot: true } }}>
+	<BubbleMap data={la_locations} lat="lat" long="long" size="sales" />
+</Story>
+
+<Story name="Loading" parameters={{ chromatic: { disableSnapshot: true } }}>
+	<BubbleMap data={slow_la_locations} lat="lat" long="long" size="sales" />
+</Story>
